services/shared/requirements.txt
opentelemetry-api>=1.20,<2
opentelemetry-sdk>=1.20,<2
opentelemetry-exporter-otlp-proto-grpc>=1.20,<2
opentelemetry-instrumentation-fastapi>=0.41b0,<1
opentelemetry-instrumentation-httpx>=0.41b0,<1
opentelemetry-instrumentation-botocore>=0.41b0,<1
boto3>=1.28,<2
pydantic>=2.0,<3
